Rename: 프로필 관련 페이지 위치 변경
